email-assistant-backend/requirements.txt
fastapi
uvicorn
python-dotenv
anthropic
openai